Type conversion fixed

+++ Common/Source/Parser.cpp
@@ -158,7 +158,7 @@ size_t NMEAParser::ValidateAndExtract(const TCHAR *src, TCHAR *dst, size_t dstsz
 {
   int len = _tcslen(src);
 
-  if (len <= 6 || len >= dstsz)
+  if (len <= 6 || len >= (int) dstsz)
     return 0;
   if (!NMEAChecksum(src))
     return 0;
@@ -382,7 +382,7 @@ BOOL NMEAParser::GSA(TCHAR *String, TCHAR **params, size_t nparams, NMEA_INFO *G
   // but 1st item in string is not passed, so start at item 3
   for (int i = 0; i < MAXSATELLITES; i++)
   {
-    if (3+i<nparams) {
+    if (3+i < (int) nparams) {
       GPS_INFO->SatelliteIDs[i] = (int)(StrToDouble(params[2+i], NULL)); // 2 because params is 0-index
       if (GPS_INFO->SatelliteIDs[i] > 0)
 	iSatelliteCount ++;
